+++ reproc/test/environment.c
@@ -20,7 +20,7 @@ int main(void)
 
   char *output = NULL;
   reproc_sink sink = reproc_sink_string(&output);
-  r = reproc_drain(process, sink, sink, REPROC_INFINITE);
+  r = reproc_drain(process, sink, sink);
   assert(r == 0);
   assert(output != NULL);